+++ src/pouch-db.ts
@@ -82,6 +82,10 @@ export function isLevelDown(adapter: any) {
     }
 }
 
+
+const validCouchDBStringRegexStr = '^[a-z][_$a-z0-9]*$';
+const validCouchDBStringRegex = new RegExp(validCouchDBStringRegexStr);
+
 /**
  * validates that a given string is ok to be used with couchdb-collection-names
  * @link https://wiki.apache.org/couchdb/HTTP_database_API
@@ -104,11 +108,9 @@ export function validateCouchDBString(name: string): true {
     }
 
 
-    const regStr = '^[a-z][_$a-z0-9]*$';
-    const reg = new RegExp(regStr);
-    if (!name.match(reg)) {
+    if (!name.match(validCouchDBStringRegex)) {
         throw newRxError('UT2', {
-            regex: regStr,
+            regex: validCouchDBStringRegexStr,
             givenName: name,
         });
     }