+++ src/rx-query.ts
@@ -42,9 +42,10 @@ import type {
 import {
     createRxDocuments
 } from './rx-document-prototype-merge';
-import { RxChangeEvent } from './rx-change-event';
+import type { RxChangeEvent } from './rx-change-event';
 import { calculateNewResults } from './event-reduce';
 import { PreparedQuery } from './rx-storate.interface';
+import { triggerCacheReplacement } from './query-cache';
 
 let _queryCount = 0;
 const newQueryID = function (): number {
@@ -59,13 +60,19 @@ export class RxQueryBase<
     public id: number = newQueryID();
 
     /**
-     * counts how often the execution on the whole db was done
-     * (used for tests and debugging)
+     * Some stats then are used for debugging and cache replacement policies
      */
     public _execOverDatabaseCount: number = 0;
+    public _creationTime = now();
+    public _lastEnsureEqual = 0;
 
     // used by some plugins
     public other: any = {};
+
+    public uncached = false;
+
+    // used to count the subscribers to the query
+    public refCount$ = new BehaviorSubject(null);
 
     constructor(
         public op: RxQueryOP,
@@ -120,7 +127,10 @@ export class RxQueryBase<
                 // tslint:disable-next-line
                 merge(
                     results$,
-                    changeEvents$
+                    changeEvents$,
+                    this.refCount$.pipe(
+                        filter(() => false)
+                    )
                 ) as any;
         }
         return this._$ as any;
@@ -155,7 +165,7 @@ export class RxQueryBase<
      * - Emit the new result-set when an RxChangeEvent comes in
      * - Do not emit anything before the first result-set was created (no null)
      */
-    private _$?: BehaviorSubject<RxQueryResult>;
+    public _$?: BehaviorSubject<RxQueryResult>;
 
     /**
      * set the new result-data as result-docs of the query
@@ -424,6 +434,8 @@ export function createRxQuery(
 
     runPluginHooks('createRxQuery', ret);
 
+    triggerCacheReplacement(collection);
+
     return ret;
 }
 
@@ -459,6 +471,7 @@ function _ensureEqual(rxQuery: RxQueryBase): Promise<boolean> {
  * @return true if results have changed
  */
 function __ensureEqual(rxQuery: RxQueryBase): Promise<boolean> | boolean {
+    rxQuery._lastEnsureEqual = now();
     if (rxQuery.collection.database.destroyed) return false; // db is closed
     if (_isResultsInSync(rxQuery)) return false; // nothing happend